refactor: move connection resource & data source to a separate package (#530)
This is broken out of #518 

Moving the connection resource and data source to a separate package
makes it easier to ensure that connection-related code remains isolated
from the provider implementation and other resources. As part of this
move, unit tests of private functions related to connection speed are
split out of the connection resource into a separate file; this enables
moving the acceptance tests into a separate package while minimizing
code changes.
