+++ equinix/resource_metal_vrf_acc_test.go
@@ -15,6 +15,10 @@ import (
 	"github.com/packethost/packngo"
 )
 
+const (
+	metalDedicatedConnIDEnvVar = "TF_ACC_METAL_DEDICATED_CONNECTION_ID"
+)
+
 func init() {
 	resource.AddTestSweepers("equinix_metal_vrf", &resource.Sweeper{
 		Name: "equinix_metal_vrf",
